+++ packages/core-js-compat/src/mapping.mjs
@@ -86,6 +86,7 @@ export default {
     [106, '1.25'],
     [107, '1.26'],
     [108, '1.27'],
+    [109, '1.28'],
   ],
   // https://github.com/Kilian/electron-to-chromium/blob/master/chromium-versions.js
   // Maybe also required to handle used Node versions?